requirements.txt
Django==5.2.8
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
psycopg[binary]==3.2.1
pytest==9.0.1
pytest-django==4.11.1
django-cors-headers==4.4.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
numpy>=1.24.0

